backend/pyproject.toml
[project]
name = "shd-casting-backend"
version = "0.1.0"
description = "SHD Casting official website backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy>=2.0",
    "pymysql>=1.1",
    "cryptography>=43",
    "alembic>=1.13",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0",
    "python-multipart>=0.0.6",
    "aiosmtplib>=3.0",
    "email-validator>=2.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]
